tender_generator_skill/requirements.txt
openai>=1.0.0
markdown
python-docx
beautifulsoup4
python-dotenv